+++ data-manager/api/src/main/java/org/orbisgis/orbisdata/datamanager/api/dataset/IDataSet.java
@@ -111,4 +111,8 @@ public interface IDataSet<T> extends Iterable<T> {
      * @return A filtered {@link IDataSet}
      */
     IDataSet<?> filter(String filter);
+
+    @Override
+    @NotNull
+    String toString();
 }