+++ data-manager/api/src/main/java/org/orbisgis/orbisdata/datamanager/api/dataset/ISummary.java
@@ -54,4 +54,8 @@ public interface ISummary {
      */
     @Nullable
     Object asType(@NotNull Class<?> clazz);
+
+    @Override
+    @NotNull
+    String toString();
 }