+++ data-manager/api/src/main/java/org/orbisgis/orbisdata/datamanager/api/datasource/IDataSource.java
@@ -72,6 +72,19 @@ public interface IDataSource<T> {
     @Nullable
     IDataSourceLocation getLocation();
 
+    /**
+     * Convert the current object into another with the given class.
+     *
+     * @param clazz New class of the result.
+     * @return The current object into an other class.
+     */
+    @Nullable
+    Object asType(@NotNull Class<?> clazz);
+
+    @Override
+    @NotNull
+    String toString();
+
 
     /* ********************** */
     /*      Load methods      */