+++ data-manager/jdbc/src/main/java/org/orbisgis/orbisdata/datamanager/jdbc/h2gis/H2GIS.java
@@ -269,4 +269,10 @@ public class H2GIS extends JdbcDataSource {
             return false;
         }
     }
+
+    @Nullable
+    @Override
+    public Object asType(@NotNull Class<?> clazz) {
+        return null;
+    }
 }