+++ data-manager/jdbc/src/main/java/org/orbisgis/orbisdata/datamanager/jdbc/postgis/POSTGIS.java
@@ -212,4 +212,10 @@ public class POSTGIS extends JdbcDataSource {
             return false;
         }
     }
+
+    @Nullable
+    @Override
+    public Object asType(@NotNull Class<?> clazz) {
+        return null;
+    }
 }